==> notebooks_ETL/01_bastion_analytics_header_concat_bronze.ipynb <==
#### Concatenating all header files

import pandas as pd

# dictionary of the lists of files from each year
files = {
    2018: [
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2018\header\ams__header_2018__202001290000_part_0.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2018\header\ams__header_2018__202001290000_part_1.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2018\header\ams__header_2018__202001290000_part_2.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2018\header\ams__header_2018__202001290000_part_3.csv",
    ],
    2019: [
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2019\header\ams__header_2019__202001080000_part_0.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2019\header\ams__header_2019__202001080000_part_1.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2019\header\ams__header_2019__202001080000_part_2.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2019\header\ams__header_2019__202001080000_part_3.csv",
    ],
    2020: [
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2020\header\ams__header_2020__202009291500_part_0.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2020\header\ams__header_2020__202009291500_part_1.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2020\header\ams__header_2020__202009291500_part_2.csv"
    ]
}

##### Sample data in each part of header

# assigns first 5 rows of each header parts into dataframe
sample_dfs = {}

for year, file_list in files.items():
    for f in file_list:
        df = pd.read_csv(f, nrows=5)
        sample_dfs[f] = df

# eda
year, file_list = next(iter(files.items()))

eda_df = pd.read_csv(file_list[0], nrows=5)
eda_df.info()

# displays sample of each header parts
for name, df in sample_dfs.items():
    print(f"\n{name}")
    display(df)
# if we concat them, we have to remove the column for part1, part2, and part3


# sees all columns
pd.set_option('display.max_columns', None)

##### Concatenating

# combined header file
output = r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\ams_header_combined.csv"
chunk_size = 500000

##### Time took for concatenating the files: 9m 49.1s

# new updated chunking for all years
first = True

for year, file_list in files.items():
    for f in file_list:
        for chunk in pd.read_csv(f, chunksize=chunk_size):
            
            # add the year column
            chunk["source_year"] = year

            # write chunk to output
            chunk.to_csv(
                output,
                mode="a",
                index=False,
                header=first
            )
            first = False
