==> notebooks_ETL/02_bastion_analytics_container_concat_bronze.ipynb <==
#### Concatenating all container files

import pandas as pd

# All files to be concatenated
files = {
    2018: [
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2018\container\ams__container_2018__202001290000_part_0.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2018\container\ams__container_2018__202001290000_part_1.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2018\container\ams__container_2018__202001290000_part_2.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2018\container\ams__container_2018__202001290000_part_3.csv",
    ],
    2019: [
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2019\container\ams__container_2019__202001080000_part_0.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2019\container\ams__container_2019__202001080000_part_1.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2019\container\ams__container_2019__202001080000_part_2.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2019\container\ams__container_2019__202001080000_part_3.csv",
    ],
    2020: [
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2020\container\ams__container_2020__202009291500_part_0.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2020\container\ams__container_2020__202009291500_part_1.csv",
        r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\2020\container\ams__container_2020__202009291500_part_2.csv",
    ]
}

##### Concate the files

##### Time took: 5m 54.3s

# variable to get only first header in part0
first = True
chunk_size = 500000
output_path = r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\ams_container_combined.csv"

for year, file_list in files.items():
    for f in file_list:
        for chunk in pd.read_csv(f, chunksize=chunk_size):
            
            # add the year column
            chunk["source_year"] = year

            # write chunk to output
            chunk.to_csv(
                output_path,
                mode="a",
                index=False,
                header=first
            )
            first = False
