==> notebooks_ETL/04_bastion_analytics_container_clean_silver.ipynb <==
#### Clean and normalize container csvs from all years

import numpy as np
import pandas as pd
import os

pd.set_option('display.max_columns', None)

container_sample = pd.read_csv(r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\ams_container_combined.csv", nrows=100)
container_sample.head(0)

#### Create a new colume for volume by multiplying the dimensions
##### The length, height, and width are measured in mm

input_path = r"..\data\bronze_raw\CBP AMS Shipping Data\raw-layer\ams_container_combined.csv"
output_path = r"..\data\silver_cleaned\ams_container_cleaned.csv"

chunk_size = 1000000

##### New column Volume (meters cubed) created from multiplying the dimensions (mm)

first = True
for chunk in pd.read_csv(input_path, chunksize=chunk_size):
    # Creates new column for Volume in cubic meters, rounds to nearest thousandth
    chunk["Volume (meters cubed)"] = (chunk["container_length"] * chunk["container_width"] * chunk["container_height"] / 1000000000).round(3)
    chunk.to_csv(output_path, mode = "w" if first else "a", index = False, header = first)
    first = False

sample_cleaned = pd.read_csv(output_path, nrows=100)
sample_cleaned.head(10)

#### Drop unneeded columns

drop_columns = [
    "seal_number_1",
    "seal_number_2",
    "equipment_description_code",
    "container_type",
    "load_status",
    "type_of_service",
]

file_path = r"..\data\silver_cleaned\ams_container_cleaned.csv"
temp_path = r"..\data\silver_cleaned\silver_container_temp.csv"
chunk_size = 1000000

first = True
for chunk in pd.read_csv(file_path, chunksize = chunk_size):
    # Keep only existing columns to avoid errors
    existing_drop_columns = [c for c in drop_columns if c in chunk.columns]
    
    chunk = chunk.drop(columns = existing_drop_columns)
    
    chunk.to_csv(temp_path, mode = "w" if first else "a", index = False, header = first)
    first = False

# replaces original with new dropped columns file
os.replace(temp_path, file_path)

container_sample = pd.read_csv(file_path, nrows=100)
container_sample.head(5)
