==> notebooks_ETL/05_bastion_analytics_lookup_tables.ipynb <==
#### Creating the lookup tables

import numpy as np
import pandas as pd

file_path = r"..\data\silver_cleaned\ams_header_cleaned.csv"
country_output_path = r"..\data\lookup_tables\countries_lookup.csv"
ports_output_path = r"..\data\lookup_tables\ports_lookup.csv"

##### Country Code Lookup Table

header = pd.read_csv(file_path, usecols=['vessel_country_code', 'port_of_unlading'])

# create ports lookup table
countries = (header[['vessel_country_code']].drop_duplicates().sort_values('vessel_country_code').reset_index(drop=True))

# this creates a country_id starting from 1
countries['country_id'] = countries.index + 1
countries = countries.rename(columns={
    'vessel_country_code': 'country_code'
})

# reorders the columns, had an issue with bulk inserting in SQL where country code was first
countries = countries[['country_id','country_code']]
countries.to_csv(country_output_path, index=False)

##### Ports Lookup Table

ports = (header[['port_of_unlading']].drop_duplicates().sort_values('port_of_unlading').reset_index(drop=True)
)

# ports id starting from 1
ports['port_id'] = ports.index + 1
ports = ports.rename(columns={
    'port_of_unlading': 'port_name'
})

# orders columns
ports = ports[['port_id','port_name']]
ports.to_csv(ports_output_path, index=False)
